Changed how HMS and DMS sets its value; fmt_angle should use b
HMS and DMS: all properties should use self.hms or self.dms so that there is
only one place where is access the angle (__sethms, __setdms).

fmt_angle is passed ``b`` but doesn't use it in call to deci2sexa. Found
an edge case where this was producing incorrect sign.

Added Angle._b. Set to False here and True in DeltaAngle. This is then used
by method that need it.

Changed docstring of DeltaAngle to reflect latest code and made it pass